Fast_API_Celine/mlflow_salary/requirements.txt
fastapi 
uvicorn[standard]
pydantic 
pandas 
gunicorn 
openpyxl 
mlflow 
boto3 
python-multipart
Pillow
numpy

